Remove bifurcating-only Sackin index and rename generalized to standard
The Sackin index formula naturally supports polytomies without any
modification, so there's no need for separate bifurcating and generalized
versions. Removed the bifurcating-only version and renamed
alifestd_mark_sackin_index_generalized_asexual to
alifestd_mark_sackin_index_asexual.
