+++ hstrat/_auxiliary_lib/__init__.pyi
@@ -239,9 +239,6 @@ from ._alifestd_mark_roots_polars import alifestd_mark_roots_polars
 from ._alifestd_mark_sackin_index_asexual import (
     alifestd_mark_sackin_index_asexual,
 )
-from ._alifestd_mark_sackin_index_generalized_asexual import (
-    alifestd_mark_sackin_index_generalized_asexual,
-)
 from ._alifestd_mark_sister_asexual import alifestd_mark_sister_asexual
 from ._alifestd_mask_descendants_asexual import (
     alifestd_mask_descendants_asexual,
@@ -556,7 +553,6 @@ __all__ = [
     "alifestd_mark_roots",
     "alifestd_mark_roots_polars",
     "alifestd_mark_sackin_index_asexual",
-    "alifestd_mark_sackin_index_generalized_asexual",
     "alifestd_mark_sister_asexual",
     "alifestd_mask_descendants_asexual",
     "alifestd_mask_monomorphic_clades_asexual",
